First working version with Data Tracing module (currently in http modules)
