+++ conf/nginx.conf
@@ -1,6 +1,7 @@
 
 #user  nobody;
-worker_processes  1;
+daemon off;
+master_process off;
 
 #error_log  logs/error.log;
 #error_log  logs/error.log  notice;
@@ -43,6 +44,7 @@ http {
         location / {
             root   html;
             index  index.html index.htm;
+	    data_tracing;
         }
 
         #error_page  404              /404.html;
